+++ curator_service/app/api/spotify_authenticate.py
