+++ curator_service/app/models/curator.py
