+++ ui/README.md
@@ -1,4 +1,4 @@
-# frontend_service
+# ui
 
 This template should help get you started developing with Vue 3 in Vite.
