+++ ui/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "frontend_service",
+  "name": "ui",
   "version": "0.0.0",
   "private": true,
   "type": "module",
