+++ ui/src/components/Ping.vue
