+++ ui/src/components/Slider.vue
