+++ ui/src/views/Curator.vue
